\name{aids.net}
\title{aids blogs data as a ``network" object}
\alias{aids}
\alias{aids.net}
\description{Network of citations among blogs related to AIDS, patients, and their support networks, collected by Gopal, over a three-day period in August 2005. 
A directed graph representing
the pattern of citation among 146 unique blogs related to AIDS, patients,
and their support networks, collected by Gopal over a randomly selected
three-day period in August 2005.  Vertices correspond to blogs.  A directed
edge from one blog to another indicates that the former had a link to the
latter in their web page (more specifically, the former refers to the latter
in their so-called 'blogroll').
}
\usage{
 data(aids)
}
\source{http://math.bu.edu/people/kolaczyk/datasets/AIDSBlog.zip
}
\seealso{network, plot.network, VBLPCM}
\references{
  S. Gopal, "The evolving social geography of blogs," in Societies
  and CIties in the Age of Instant Access, H. Miller, Ed.   Berlin:Springer,
  2007, pp. 275-294
}
\keyword{datasets}
