\name{print.vblpcm}
\alias{print.vblpcm}
\title{print the fitted vblpcm object
}
\description{Print a vblpcm object. 
}
\usage{
\method{print}{vblpcm}(x, \dots)
}
\arguments{
  \item{x}{The fitted values; output from vblpcmfit()
}
  \item{\dots}{optional arguments to be passed to lower level functions
}
}
\author{Michael Salter-Townshend
}

\seealso{latentnet::print.ergmm
}
