\name{simulated.network}
\title{simulated.network}
\alias{Y}
\description{adjacency matrix simulated from the latent position cluster model with 3 well separated groups
}
\usage{
 data(simulated.network)
}
\source{Michael Salter-Townshend
}
\seealso{network, plot.network, VBLPCM}
\keyword{datasets}
