\name{summary.vblpcm}
\alias{summary.vblpcm}
\title{summary of a fitted vblpcm object. 
}
\description{Summarise the output of a call to either vblpcmstart or vblpcmfit. 
}
\usage{
\method{summary}{vblpcm}(object, \dots)
}
\arguments{
  \item{object}{The fitted values; output from vblpcmstart() or vblpcmfit()
}
  \item{\dots}{optional arguments to be passed to lower level functions
}
}
\author{Michael Salter-Townshend
}

\seealso{latentnet::summary.ergmm
}
