\name{vblpcmgroups}
\alias{vblpcmgroups}
\title{list the maximum VB a-posteriori group memberships. 
}
\description{Prints to screen the most likely a-posteriori membership of each node. 
}
\usage{vblpcmgroups(v.params, colours)
}
\arguments{
  \item{v.params}{The fitted values; output from vblpcmfit()
}
  \item{colours}{The colours to be used.
}
}
\value{Prints to screen of the most probable group membership for each node. 
}
\author{Michael Salter-Townshend
}
